backend/requirements.txt
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
attrs==25.4.0
backports.zstd==1.2.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
clerk-backend-api==4.2.0
click==8.3.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
fastapi==0.104.1
fonttools==4.61.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.0
mypy==1.7.1
mypy_extensions==1.1.0
narwhals==2.14.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pluggy==1.6.0
propcache==0.4.1
pycares==4.11.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
yarl==1.22.0
anthropic>=0.40.0
sse-starlette==2.1.0
